rss_collector/pyproject.toml
[project]
name = "rss-collector"
version = "0.1.0"
description = "RSS Financial Intelligence Collector for OpenClaw"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.0",
    "uvicorn==0.35.0",
    "feedparser>=6.0.0",
    "newspaper4k>=0.9.0",
    "motor>=3.3.0",
    "httpx>=0.28.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.11,<2.12",
]

[dependency-groups]
dev = []
